added console color and exe now can parse through multiple files
