+++ .gitignore
@@ -10,7 +10,7 @@ urls.txt
 *.user
 *.userosscache
 *.sln.docstates
-
+CheckLinkCLI2/CheckLinkCLI2/index2.html
 [Dd]ebug/
 [Dd]ebugPublic/
 [Rr]elease/
