+++ CheckLinkCLI2/CheckLinkCLI2/Program.cs
@@ -5,168 +5,107 @@ using System.Security.Cryptography.X509Certificates;
 using System.Threading;
 using System.Collections;
 using System.Collections.Concurrent;
+using System.Net;
+using System.Net.Http;
+using System.Threading.Tasks;
+using System.Linq;
 
 namespace CheckLinkCLI2
 {
     public class Program
     {
-        static ConcurrentDictionary<int, int> items = new ConcurrentDictionary<int, int>();
+        #region Debug or Release
+        public static bool IsDebug()
+        {
+#if (DEBUG)
+            return true;
+#else
+            return false;
+#endif
+        }
 
-        //static void Main(string[] args)
+        #endregion
+
+        private static readonly string linkFile = @"D:\Documents\Seneca College\OSD600\check-link\check-link\CheckLinkCLI2\CheckLinkCLI2\urls.txt";
+        public static readonly List<string> version = new List<string>() { "v", "-v", "version", "--version" };
+        public static Dictionary<string, List<string>> CommandLineOptions = new Dictionary<string, List<string>>()
+        {
+            {"version",version }
+        };
         //{
-        //    /// <summary>
-        //    /// Basics on collections
-        //    /// </summary>
-
-        //    List<String> customer = new List<string>();
-        //    customer.Add("Kim");
-        //    customer.Add("John");
-        //    customer.Add("Tim");
-        //    Console.WriteLine(customer.Count);
-
-        //    foreach (var item in customer)
-        //    {
-        //        Console.WriteLine(item);
-        //    }
-
-        //    Console.WriteLine(customer[1]);
-
-        //    /// <summary>
-        //    /// Intro to Dictionary
-        //    /// </summary>
-        //    ///
-        //    Dictionary<string, string> config = new Dictionary<string, string>();
-
-        //    config.Add("resolution", "1920x1080");
-        //    config.Add("title", "MyWebsite");
-
-        //    /// This is not possible
-        //    //config.Add("title", "MyWebsitea");
-
-        //    Console.WriteLine(config["title"]);
-
-        //    foreach (var setting in config)
-        //    {
-        //        Console.WriteLine(setting.Value);
-
-        //    }
-
-        //    /// <summary>
-        //    /// Running array list and the concept of Boxing/Unboxing
-        //    /// </summary>
-        //    ///
-        //    ArrayList list = new ArrayList();
-
-        //    // this is possible because of boxing; which is converting any type to a regular C# object
-        //    list.Add("some string");
-        //    string s = (string)list[0];
-
-        //    /// <summary>
-        //    /// Difference between HashTable collections
-        //    /// </summary>
-        //    ///
-        //    Hashtable table = new Hashtable();
-
-        //    table.Add("title", "MyWebsite");
-
-        //    var j = (string)table["title"];
-        //    Console.WriteLine($"This is a hashtable value {j}");
-
-        //    /// <summary>
-        //    /// Learning a Concurrent Dictionary
-        //    /// </summary>
-        //    ///
-
-
-        //    Thread thread1 = new Thread(new ThreadStart(AddItem));
-        //    Thread thread2 = new Thread(new ThreadStart(AddItem));
-
-        //    //thread1.Start();
-        //    //thread2.Start();
-
-        //    bool[] preload = new bool[3] { true, false, true};
-
-        //    BitArray enemyGrid = new BitArray(preload);
-
-        //    //enemyGrid[0] = false;
-        //    //enemyGrid[1] = true;
-        //    //enemyGrid.Set(2,false);
-        //    //enemyGrid
-
-        //    foreach (var item in enemyGrid)
-        //    {
-        //        Console.WriteLine(item);
-        //    }
-
-        //    /// <summary>
-        //    /// Learning a Concurrent Dictionary and Tuples
-        //    /// </summary>
-        //    ///
-
-        //    Tuple<int, string, bool> myTuple = new Tuple<int, string, bool>(1, "hello", true);
-
-        //    Console.WriteLine(myTuple.Item2);
-
-        //    /// <summary>
-        //    /// Learning about Stack
-        //    /// </summary>
-        //    ///
-
-        //    Stack<string> pancakes = new Stack<string>();
-
-        //    pancakes.Push("first pancake made");
-        //    pancakes.Push("second pancake made");
-        //    pancakes.Push("third pancake made");
-
-        //    foreach (var item in pancakes)
-        //    {
-        //        Console.WriteLine(item);
-        //    }
-
-        //    Console.WriteLine(pancakes.Pop());
-        //    Console.WriteLine(pancakes.Peek()); // Stack.Peek() looks at the last item without removing it
-        //    Console.WriteLine(pancakes.Peek());
-
-        //    /// <summary>
-        //    /// Learning about Queue
-        //    /// </summary>
-        //    ///
-
-        //    Queue<int> myQueue = new Queue<int>();
-
-        //    myQueue.Enqueue(1);
-        //    myQueue.Enqueue(2);
-        //    myQueue.Enqueue(3);
-
-        //    foreach (var queue in myQueue)
-        //    {
-        //        Console.WriteLine(queue);
-        //    }
-
-        //    Console.WriteLine(myQueue.Dequeue());
-        //    Console.WriteLine(myQueue.Peek()); // Queue.Peek() does the same as Stack.Peek()
-
-        //    /// <summary>
-        //    /// Learning about HashSet
-        //    /// </summary>
-        //    ///
-
-        //    var myHash = new HashSet<string>();
-
-        //    myHash.Add("hello");
-        //    myHash.Add("hello");
-
-        //    string[] l = new string[] { "hello" };
-
-        //    Console.WriteLine(myHash.Count);
-        //    Console.WriteLine(myHash.Overlaps(l)); //Overlaps allows us to see and compare whether we have the same data
+        //     { "version", ["v","-v","version","--version"] },
 
         //}
 
-        public static void AddItem()
+
+        public static void Main(string[] args)
         {
-            items.TryAdd(1, 2);
-            Console.WriteLine(items.Count);
+            FileReader FileReader = new FileReader();
+            WebLinkChecker LinkChecker = new WebLinkChecker();
+
+            if (IsDebug())
+            {
+                //TODO: Search function that checks only one or few links on-demand
+
+                //Type type = typeof(ConsoleColor);
+                //Console.ForegroundColor = ConsoleColor.White;
+                //foreach (var name in Enum.GetNames(type))
+                //{
+                //    Console.BackgroundColor = (ConsoleColor)Enum.Parse(type, name);
+                //    Console.WriteLine(name);
+                //}
+                //Console.BackgroundColor = ConsoleColor.Black;
+                //foreach (var name in Enum.GetNames(type))
+                //{
+                //    Console.ForegroundColor = (ConsoleColor)Enum.Parse(type, name);
+                //    Console.WriteLine(name);
+                //}
+
+                var links = FileReader.ExtractLinks(linkFile);
+                foreach (var link in links)
+                {
+                    LinkChecker.GetAllEndPointWithUri(link);
+                }
+            }
+
+            else
+            {
+                #region Command line options
+                if (version.Contains(args[0]))
+                {
+                    Console.WriteLine("Application Name: CheckLinkCLI2 \n" +
+                        "Release: 0.1");
+                }
+                #endregion
+
+                if (args.Length == 0)
+                {
+                    Console.WriteLine("Please provide file name with links as an argument...");
+                    Console.WriteLine("For example: CheckLinksCLI2 file_name.txt");
+
+                }
+
+                else
+                {
+                    foreach (var file in args)
+                    {
+                        var links = FileReader.ExtractLinks(file);
+                        foreach (var link in links)
+                        {
+                            LinkChecker.GetAllEndPointWithUri(link);
+                        }
+                    }
+
+                }
+            }
+
+
         }
+
+
+
+
+
     }
+
 }
